LayoutTests/editing/pasteboard/copy-null-characters-expected.txt

If there are NULL characters in text nodes, they should not be copied to the clipboard.  This test requires DumpRenderTree.
| "
"
| <div>
|   id="outerSource"
|   <p>
|     <b>
|       "bold"
|   "
  "
|   "
  "
|   <p>
|     style="color: green"
|     "green"
|   "
"
| "
"
| <div>
|   contenteditable="true"
|   id="destination-rich-text"
|   <p>
|     <b>
|       "bold"
|   <div>
|     id="source"
|     "Copy paste me"
|   <span>
|     style="color: green;"
|     "green"
|   "Copy paste me"
| "
"
| <textarea>
|   id="destination-plain-text"
|   this.value="Copy paste mebold

Copy paste me
green"
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "Copy paste mebold

Copy paste me
green"
| "
"
| <div>
|   id="results"
|   "PASSED"
| "

"
